dashboard/requirements.txt
pandas
sqlalchemy
psycopg2-binary
dash
python-dotenv
plotly